fixing invalid hook call
