+++ chat-widget/src/hooks/useDebounce.ts
@@ -1,133 +1,25 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-export default function useDebounce(func: any, wait?: number) {
-    wait = wait || 0;
-    let lastArgs:any, lastThis:any, maxWait:any, result:any, timerId:any, lastCallTime:any;
-  
-    let lastInvokeTime = 0;
-    const leading = false;
-    const maxing = false;
-    const trailing = true;
-  
-    if (typeof func !== "function") {
-        throw new TypeError("Expected a function");
-    }
+import { useRef, useEffect } from "react";
 
-  
-    function invokeFunc(time:number) {
-        const args = lastArgs;
-        const thisArg = lastThis;
-  
-        lastArgs = lastThis = undefined;
-        lastInvokeTime = time;
-        result = func.apply(thisArg, args);
-        return result;
-    }
-  
-    function startTimer(pendingFunc:any, wait?:number) {
-        wait = wait || 0;
-        return setTimeout(pendingFunc, wait);
-    }
-  
-    function cancelTimer(id:any) {
-        clearTimeout(id);
-    }
-  
-    function leadingEdge(time:number) {
-        // Reset any `maxWait` timer.
-        lastInvokeTime = time;
-        // Start the timer for the trailing edge.
-        timerId = startTimer(timerExpired, wait);
-        // Invoke the leading edge.
-        return leading ? invokeFunc(time) : result;
-    }
-  
-    function remainingWait(time:number) {
-        wait = wait || 0;
-        const timeSinceLastCall = time - lastCallTime;
-        const timeSinceLastInvoke = time - lastInvokeTime;
-        const timeWaiting = wait - timeSinceLastCall;
-  
-        return maxing
-        ? Math.min(timeWaiting, maxWait - timeSinceLastInvoke)
-        : timeWaiting;
-    }
-  
-    function shouldInvoke(time:number) {
-        wait = wait || 0;
-        const timeSinceLastCall = time - lastCallTime;
-        const timeSinceLastInvoke = time - lastInvokeTime;
-  
-        // Either this is the first call, activity has stopped and we're at the
-        // trailing edge, the system time has gone backwards and we're treating
-        // it as the trailing edge, or we've hit the `maxWait` limit.
-        return (lastCallTime === undefined || (timeSinceLastCall >= wait) ||
-        (timeSinceLastCall < 0) || (maxing && timeSinceLastInvoke >= maxWait));
-    }
-  
-    function timerExpired() {
-        const time = Date.now();
-        if (shouldInvoke(time)) {
-            return trailingEdge(time);
-        }
-        // Restart the timer.
-        timerId = startTimer(timerExpired, remainingWait(time));
-    }
-  
-    function trailingEdge(time:number) {
-        timerId = undefined;
-  
-        // Only invoke if we have `lastArgs` which means `func` has been
-        // debounced at least once.
-        if (trailing && lastArgs) {
-            return invokeFunc(time);
-        }
-        lastArgs = lastThis = undefined;
-        return result;
-    }
-  
-    function cancel() {
-        if (timerId !== undefined) {
-            cancelTimer(timerId);
-        }
-        lastInvokeTime = 0;
-        lastArgs = lastCallTime = lastThis = timerId = undefined;
-    }
-  
-    function flush() {
-        return timerId === undefined ? result : trailingEdge(Date.now());
-    }
-  
-    function pending() {
-        return timerId !== undefined;
-    }
-  
-    function debounced(this: any, ...args: any[]) {
-        const time = Date.now();
-        const isInvoking = shouldInvoke(time);
-  
-        lastArgs = args;
-        // eslint-disable-next-line @typescript-eslint/no-this-alias
-        lastThis = this;
-        lastCallTime = time;
-  
-        if (isInvoking) {
-            if(timerId === undefined) {
-                return leadingEdge(lastCallTime);
-            }
-            if (maxing) {
-                // Handle invocations in a tight loop.
-                timerId = startTimer(timerExpired, wait);
-                return invokeFunc(lastCallTime);
-            }
-        }
-        if (timerId === undefined) {
-            timerId = startTimer(timerExpired, wait);
-        }
-        return result;
-    }
-    
-    debounced.cancel = cancel;
-    debounced.flush = flush;
-    debounced.pending = pending;
-    return debounced;
+type Timer = ReturnType<typeof setTimeout>;
+type SomeFunction = (...args: unknown[]) => void;
+
+export default function useDebounce<Fn extends SomeFunction>(func: Fn, delay = 1000) {
+    const timer = useRef<Timer>();
+
+    useEffect(() => {
+        return () => {
+            if (!timer.current) return;
+            clearTimeout(timer.current);
+        };
+    }, []);
+
+    const debouncedFunction = ((...args) => {
+        const newTimer = setTimeout(() => {
+            func(...args);
+        }, delay);
+        clearTimeout(timer.current);
+        timer.current = newTimer;
+    }) as Fn;
+
+    return debouncedFunction;
 }
