+++ chat-widget/src/hooks/useWindowDimensions.ts
@@ -1,5 +1,5 @@
 import { useState, useEffect } from "react";
-import useDebounce from "./useDebounce2";
+import useDebounce from "./useDebounce";
 
 function getWindowDimensions() {
     const { innerWidth: width, innerHeight: height } = window;
@@ -12,11 +12,10 @@ function getWindowDimensions() {
 export default function useWindowDimensions(delay = 200) {
   
     const [windowDimensions, setWindowDimensions] = useState(getWindowDimensions());
+    const handleResize = () => setWindowDimensions(getWindowDimensions());
+    const debouncedHandleResize = useDebounce(handleResize, delay);
 
-    useEffect(() => {
-        const handleResize = () => setWindowDimensions(getWindowDimensions());
-        const debouncedHandleResize = useDebounce(handleResize, delay);
-        
+    useEffect(() => {        
         window.addEventListener("resize", debouncedHandleResize);
         return () => window.removeEventListener("resize", debouncedHandleResize);
     }, []);
